core: clean API visibility of pg_brick_* and lifecycle
- pg_bricks_* headers lived in packetgraph.h, moved them to
  include/brick.h
- src/brick.h and src/brick-int.h separation do not have sens anymore,
  merging them into brick-int.h
- src/bricks.c do not match headers naming => rename bricks.c to brick.c
- lifecycle headers lived in packetgraph.h moved them to
  include/lifecycle.h and fix doc
- packetgraph.h now only include other headers
