+++ src/hub.c
@@ -15,8 +15,8 @@
  * along with Butterfly.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include <packetgraph/hub.h>
-#include "brick.h"
+#include <packetgraph/packetgraph.h>
+#include "brick-int.h"
 
 struct pg_hub_state {
 	struct pg_brick brick;
