+++ src/vhost.c
@@ -31,9 +31,8 @@
 #include <linux/virtio_ring.h>
 #include <rte_virtio_net.h>
 
-#include <packetgraph/common.h>
-#include <packetgraph/vhost.h>
-#include "brick.h"
+#include <packetgraph/packetgraph.h>
+#include "brick-int.h"
 #include "packets.h"
 #include "utils/bitmask.h"
 #include "utils/mempool.h"
