+++ tests/antispoof/tests.c
@@ -23,7 +23,7 @@
 #include <packetgraph/packetgraph.h>
 #include <packetgraph/antispoof.h>
 #include "packetsgen.h"
-#include "brick.h"
+#include "brick-int.h"
 #include "packets.h"
 #include "utils/mempool.h"
 #include "utils/bitmask.h"
