+++ tests/core/test-flow.c
@@ -18,7 +18,7 @@
 #include <glib.h>
 
 #include <packetgraph/nop.h>
-#include "brick.h"
+#include "brick-int.h"
 #include "utils/bitmask.h"
 #include "utils/config.h"
 #include "tests.h"
