+++ tests/core/test-hub.c
@@ -21,7 +21,7 @@
 #include <string.h>
 #include <packetgraph/packetgraph.h>
 #include <packetgraph/hub.h>
-#include "brick.h"
+#include "brick-int.h"
 #include "packets.h"
 #include "utils/mempool.h"
 #include "utils/bitmask.h"
