+++ tests/diode/tests.c
@@ -19,7 +19,7 @@
 #include <string.h>
 #include <packetgraph/packetgraph.h>
 #include <packetgraph/diode.h>
-#include "brick.h"
+#include "brick-int.h"
 #include "packets.h"
 #include "utils/mempool.h"
 #include "utils/bitmask.h"
