+++ tests/nic/bench-nic.c
@@ -27,7 +27,7 @@
 #include <packetgraph/packetgraph.h>
 #include <packetgraph/nic.h>
 #include "packets.h"
-#include "brick.h"
+#include "brick-int.h"
 #include "utils/bench.h"
 #include "utils/mempool.h"
 #include "utils/bitmask.h"
